backend/requirements.txt
 django
 psycopg2-binary
 requests
 PyJWT
 psycopg[binary]